+++ apps/meteor/client/components/message/variants/ThreadMessage.tsx
@@ -1,6 +1,7 @@
 import { type IThreadMessage, type IThreadMainMessage, isVideoConfMessage } from '@rocket.chat/core-typings';
 import { Message, MessageLeftContainer, MessageContainer } from '@rocket.chat/fuselage';
 import { useToggle } from '@rocket.chat/fuselage-hooks';
+import { MessageAvatar } from '@rocket.chat/ui-avatar';
 import { useUserId } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
 import React, { memo, useRef } from 'react';
@@ -8,12 +9,12 @@ import React, { memo, useRef } from 'react';
 import type { MessageActionContext } from '../../../../app/ui-utils/client/lib/MessageAction';
 import { useIsMessageHighlight } from '../../../views/room/MessageList/contexts/MessageHighlightContext';
 import { useJumpToMessage } from '../../../views/room/MessageList/hooks/useJumpToMessage';
-import { useChat } from '../../../views/room/contexts/ChatContext';
+import { useUserCard } from '../../../views/room/contexts/UserCardContext';
+import Emoji from '../../Emoji';
 import IgnoredContent from '../IgnoredContent';
 import MessageHeader from '../MessageHeader';
-import MessageToolboxHolder from '../MessageToolboxHolder';
+import MessageToolbarHolder from '../MessageToolbarHolder';
 import StatusIndicators from '../StatusIndicators';
-import MessageAvatar from '../header/MessageAvatar';
 import ThreadMessageContent from './thread/ThreadMessageContent';
 
 type ThreadMessageProps = {
@@ -27,7 +28,7 @@ const ThreadMessage = ({ message, sequential, unread, showUserAvatar }: ThreadMe
 	const uid = useUserId();
 	const editing = useIsMessageHighlight(message._id);
 	const [ignored, toggleIgnoring] = useToggle((message as { ignored?: boolean }).ignored);
-	const chat = useChat();
+	const { openUserCard, triggerProps } = useUserCard();
 
 	const messageRef = useRef(null);
 
@@ -38,6 +39,8 @@ const ThreadMessage = ({ message, sequential, unread, showUserAvatar }: ThreadMe
 
 	return (
 		<Message
+			role='listitem'
+			tabIndex={0}
 			id={message._id}
 			ref={messageRef}
 			isEditing={editing}
@@ -54,14 +57,14 @@ const ThreadMessage = ({ message, sequential, unread, showUserAvatar }: ThreadMe
 			<MessageLeftContainer>
 				{!sequential && message.u.username && showUserAvatar && (
 					<MessageAvatar
-						emoji={message.emoji}
+						emoji={message.emoji ? <Emoji emojiHandle={message.emoji} fillContainer /> : undefined}
 						avatarUrl={message.avatar}
 						username={message.u.username}
 						size='x36'
-						{...(chat?.userCard && {
-							onClick: chat?.userCard.open(message.u.username),
-							style: { cursor: 'pointer' },
-						})}
+						onClick={(e) => openUserCard(e, message.u.username)}
+						style={{ cursor: 'pointer' }}
+						role='button'
+						{...triggerProps}
 					/>
 				)}
 				{sequential && <StatusIndicators message={message} />}
@@ -72,7 +75,7 @@ const ThreadMessage = ({ message, sequential, unread, showUserAvatar }: ThreadMe
 
 				{ignored ? <IgnoredContent onShowMessageIgnored={toggleIgnoring} /> : <ThreadMessageContent message={message} />}
 			</MessageContainer>
-			{!message.private && <MessageToolboxHolder message={message} context={messageContext} />}
+			{!message.private && <MessageToolbarHolder message={message} context={messageContext} />}
 		</Message>
 	);
 };