+++ apps/meteor/ee/client/omnichannel/ContactManagerInfo.js
@@ -1,8 +1,8 @@
 import { css } from '@rocket.chat/css-in-js';
+import { UserAvatar } from '@rocket.chat/ui-avatar';
 import React, { useMemo } from 'react';
 
 import { UserStatus } from '../../../client/components/UserStatus';
-import UserAvatar from '../../../client/components/avatar/UserAvatar';
 import { AsyncStatePhase } from '../../../client/hooks/useAsyncState';
 import { useEndpointData } from '../../../client/hooks/useEndpointData';
 import AgentInfoDetails from '../../../client/views/omnichannel/components/AgentInfoDetails';