+++ apps/meteor/ee/client/omnichannel/tags/TagEdit.tsx
@@ -127,11 +127,13 @@ const TagEdit = ({ tagData, currentDepartments }: TagEditProps) => {
 					</Button>
 				</ButtonGroup>
 				{_id && (
-					<ButtonGroup stretch mbs={8}>
-						<Button icon='trash' danger onClick={() => handleDeleteTag(_id)}>
-							{t('Delete')}
-						</Button>
-					</ButtonGroup>
+					<Box mbs={8}>
+						<ButtonGroup stretch>
+							<Button icon='trash' danger onClick={() => handleDeleteTag(_id)}>
+								{t('Delete')}
+							</Button>
+						</ButtonGroup>
+					</Box>
 				)}
 			</ContextualbarFooter>
 		</Contextualbar>