+++ apps/meteor/server/modules/watchers/watchers.module.ts
@@ -39,7 +39,7 @@ import {
 
 import { subscriptionFields, roomFields } from '../../../lib/publishFields';
 import type { DatabaseWatcher } from '../../database/DatabaseWatcher';
-import { broadcastMessageSentEvent } from './lib/messages';
+import { getMessageToBroadcast } from './lib/messages';
 
 type BroadcastCallback = <T extends keyof EventSignatures>(event: T, ...args: Parameters<EventSignatures[T]>) => Promise<void>;
 
@@ -64,25 +64,21 @@ export function isWatcherRunning(): boolean {
 	return watcherStarted;
 }
 
-const messageWatcher = (watcher: DatabaseWatcher, broadcast: BroadcastCallback): void => {
-	watcher.on<IMessage>(Messages.getCollectionName(), async ({ clientAction, id, data }) => {
-		switch (clientAction) {
-			case 'inserted':
-			case 'updated':
-				void broadcastMessageSentEvent({
-					id,
-					data,
-					broadcastCallback: (message) => broadcast('watch.messages', { clientAction, message }),
-				});
-				break;
-		}
-	});
-};
-
 export function initWatchers(watcher: DatabaseWatcher, broadcast: BroadcastCallback): void {
-	const dbWatchersEnabled = !dbWatchersDisabled;
-	if (dbWatchersEnabled) {
-		messageWatcher(watcher, broadcast);
+	// watch for changes on the database and broadcast them to the other instances
+	if (!dbWatchersDisabled) {
+		watcher.on<IMessage>(Messages.getCollectionName(), async ({ clientAction, id, data }) => {
+			switch (clientAction) {
+				case 'inserted':
+				case 'updated':
+					const message = await getMessageToBroadcast({ id, data });
+					if (!message) {
+						return;
+					}
+					void broadcast('watch.messages', { message });
+					break;
+			}
+		});
 	}
 
 	watcher.on<ISubscription>(Subscriptions.getCollectionName(), async ({ clientAction, id, data, diff }) => {
@@ -159,7 +155,7 @@ export function initWatchers(watcher: DatabaseWatcher, broadcast: BroadcastCallb
 
 			case 'removed': {
 				const trash = (await Subscriptions.trashFindOneById(id, {
-					projection: { u: 1, rid: 1 },
+					projection: { u: 1, rid: 1, t: 1 },
 				})) as Pick<ISubscription, 'u' | 'rid' | '_id'> | undefined;
 				const subscription = trash || { _id: id };
 				void broadcast('watch.subscriptions', { clientAction, subscription });