+++ apps/meteor/tests/e2e/fixtures/inject-initial-data.ts
@@ -58,6 +58,22 @@ export default async function injectInitialData() {
 				value: false,
 			},
 			{
+				_id: 'SAML_Custom_Default_provider',
+				value: 'test-sp',
+			},
+			{
+				_id: 'SAML_Custom_Default_issuer',
+				value: 'http://localhost:3000/_saml/metadata/test-sp',
+			},
+			{
+				_id: 'SAML_Custom_Default_entry_point',
+				value: 'http://localhost:8080/simplesaml/saml2/idp/SSOService.php',
+			},
+			{
+				_id: 'SAML_Custom_Default_idp_slo_redirect_url',
+				value: 'http://localhost:8080/simplesaml/saml2/idp/SingleLogoutService.php',
+			},
+			{
 				_id: 'Accounts_OAuth_Google',
 				value: false,
 			},