+++ apps/meteor/tests/e2e/page-objects/admin.ts
@@ -28,6 +28,34 @@ export class Admin {
 		return this.page.locator(`label >> text=Private`);
 	}
 
+	get roomNameInput(): Locator {
+		return this.page.locator('input[name="roomName"]');
+	}
+
+	get archivedLabel(): Locator {
+		return this.page.locator('label >> text=Archived');
+	}
+
+	get archivedInput(): Locator {
+		return this.page.locator('input[name="archived"]');
+	}
+
+	get favoriteLabel(): Locator {
+		return this.page.locator('label >> text=Favorite');
+	}
+
+	get favoriteInput(): Locator {
+		return this.page.locator('input[name="favorite"]');
+	}
+
+	get defaultLabel(): Locator {
+		return this.page.locator('label >> text=Default');
+	}
+
+	get defaultInput(): Locator {
+		return this.page.locator('input[name="isDefault"]');
+	}
+
 	get inputSearchUsers(): Locator {
 		return this.page.locator('input[placeholder="Search Users"]');
 	}