+++ ee/packages/ddp-client/src/types/streams.ts
@@ -9,7 +9,6 @@ import type {
 	IEmoji,
 	ICustomSound,
 	INotificationDesktop,
-	IWebdavAccount,
 	VoipEventDataSignature,
 	IUser,
 	IOmnichannelRoom,
@@ -18,12 +17,13 @@ import type {
 	IIntegrationHistory,
 	IUserDataEvent,
 	ICalendarNotification,
-	IUserStatus,
 	ILivechatInquiryRecord,
 	ILivechatAgent,
 	IImportProgress,
 	IBanner,
 	LicenseLimitKind,
+	ICustomUserStatus,
+	IWebdavAccount,
 } from '@rocket.chat/core-typings';
 import type * as UiKit from '@rocket.chat/ui-kit';
 
@@ -97,6 +97,7 @@ export interface StreamerEvents {
 							_id: string;
 							u?: Pick<IUser, '_id' | 'username' | 'name'>;
 							rid?: string;
+							t?: string;
 						},
 				  ]
 				| [
@@ -198,7 +199,7 @@ export interface StreamerEvents {
 			key: 'updateCustomUserStatus';
 			args: [
 				{
-					userStatusData: IUserStatus;
+					userStatusData: Omit<ICustomUserStatus, '_updatedAt'>;
 				},
 			];
 		},
@@ -234,8 +235,20 @@ export interface StreamerEvents {
 		},
 		{ key: 'Users:NameChanged'; args: [Pick<IUser, '_id' | 'name' | 'username'>] },
 		{ key: 'private-settings-changed'; args: ['inserted' | 'updated' | 'removed' | 'changed', ISetting] },
-		{ key: 'deleteCustomUserStatus'; args: [{ userStatusData: unknown }] },
-		{ key: 'user-status'; args: [[IUser['_id'], IUser['username'], 0 | 1 | 2 | 3, IUser['statusText'], IUser['name'], IUser['roles']]] },
+		{ key: 'deleteCustomUserStatus'; args: [{ userStatusData: Omit<ICustomUserStatus, '_updatedAt'> }] },
+		{
+			key: 'user-status';
+			args: [
+				[
+					uid: IUser['_id'],
+					username: IUser['username'],
+					status: 0 | 1 | 2 | 3,
+					statusText: IUser['statusText'],
+					name: IUser['name'],
+					roles: IUser['roles'],
+				],
+			];
+		},
 		{
 			key: 'Users:Deleted';
 			args: [